Namespaced the PerformanceReport representers
