+++ app/representers/api/v1/performance_report/exports_representer.rb
@@ -1,0 +1,7 @@
+module Api::V1::PerformanceReport
+  class ExportsRepresenter < Roar::Decorator
+    include Representable::JSON::Collection
+
+    items extend: ExportRepresenter
+  end
+end
