+++ app/representers/api/v1/performance_report/representer.rb
@@ -1,0 +1,7 @@
+module Api::V1::PerformanceReport
+  class Representer < Roar::Decorator
+    include Representable::JSON::Collection
+
+    items extend: PeriodRepresenter
+  end
+end
