--- app/representers/api/v1/performance_report_exports_representer.rb
@@ -1,7 +1,0 @@
-module Api::V1
-  class PerformanceReportExportsRepresenter < Roar::Decorator
-    include Representable::JSON::Collection
-
-    items extend: PerformanceReportExportRepresenter
-  end
-end
